image_segmentation.py
from transformers import SegformerFeatureExtractor, SegformerForSemanticSegmentation
from PIL import Image
import torch

def load_model():
    # Load the feature extractor and model from Hugging Face
    feature_extractor = SegformerFeatureExtractor.from_pretrained("nvidia/segformer-b0-finetuned-cityscapes-1024-1024")
    model = SegformerForSemanticSegmentation.from_pretrained("nvidia/segformer-b0-finetuned-cityscapes-1024-1024")
    return feature_extractor, model

def preprocess_image(image_path, feature_extractor):
    # Load the image
    image = Image.open(image_path).convert("RGB")
    # Preprocess the image
    inputs = feature_extractor(images=image, return_tensors="pt")
    return inputs

def perform_segmentation(inputs, model):
    # Perform segmentation
    outputs = model(**inputs)
    # The model outputs logits which we need to convert to pixel-wise labels
    logits = outputs.logits
    # Apply softmax to calculate probabilities
    probs = torch.nn.functional.softmax(logits, dim=1)
    # Take the argmax across channels to get the label for each pixel
    predictions = torch.argmax(probs, dim=1)
    return predictions.squeeze().cpu().numpy()

def process_image(image_path):
    feature_extractor, model = load_model()
    inputs = preprocess_image(image_path, feature_extractor)
    segmentation_map = perform_segmentation(inputs, model)
    # Here, segmentation_map is a 2D numpy array where each value represents a class label for each pixel
    # You can further process the segmentation_map as needed, e.g., visualize it or save it to a file
    print("Segmentation completed.")
    # For demonstration, let's just return a placeholder
    return "Segmentation result placeholder"

# Example usage
if __name__ == "__main__":
    image_path = "path/to/your/image.jpg"
    process_image(image_path)