requirements.txt
Flask
transformers
flask-cors
tensorflow
keras
torch 
pillow